refactor(conformance): move the run loop into a shared hook
`ConformancePanel` owned the whole conformance run inline: per-suite state,
the four route calls, the OAuth-conformance popup with its
postMessage/BroadcastChannel callback, per-suite score derivation, and the
pooled headline. score.mcpjam.com runs the same four suites, and a second
copy would drift on exactly the parts that must not — what "done" means per
suite, which transports can run which suite, and the fact that the headline
pools COUNTS rather than averaging suite scores.

This is a boundary shift, not a rewrite: the logic moves verbatim into
`useConformanceRun` and the panel becomes its first consumer. Rendering
stays where it is. The panel's 27 tests pass unchanged, which is the claim
being made here.

One thing is genuinely new, and inert until something sets it:
`deferOAuthAuthorization` parks a run at the OAuth authorization step
(`status: "needs-authorization"`) instead of opening the popup, handing the
decision to the caller via `authorizeOAuth()`. The panel passes false and
behaves exactly as before — an operator who pressed "Run" asked for the
whole run. A public landing page is a different setting, where an
unrequested popup is not something a visitor asked for.
