+++ mcpjam-inspector/client/src/components/conformance/ConformancePanel.tsx
@@ -1,4 +1,4 @@
-import { useState, useCallback, useEffect, useMemo, useRef } from "react";
+import { useState, useMemo } from "react";
 import { Button } from "@mcpjam/design-system/button";
 import {
   Select,
@@ -20,11 +20,8 @@ import {
 } from "lucide-react";
 import type { ServerWithName } from "@/hooks/use-app-state";
 import type {
-  MCPConformanceResult,
-  MCPAppsConformanceResult,
   MCPCheckResult,
   MCPAppsCheckResult,
-  MCPTasksConformanceResult,
   MCPTasksCheckResult,
   OAuthConformanceStepResult,
 } from "@mcpjam/sdk";
@@ -33,7 +30,6 @@ import type {
 // break the Vite browser bundle.
 import {
   APPS_CHECK_CATALOG,
-  canRunConformance,
   CHECK_ERAS,
   CONFORMANCE_CHECK_METADATA,
   describeConformanceScore,
@@ -41,33 +37,17 @@ import {
   MCP_CHECK_IDS,
   MCP_PROTOCOL_VERSIONS,
   MCP_TASKS_CHECK_IDS,
-  pooledConformanceScore,
   PROTOCOL_CHECK_CATALOG,
   PROTOCOL_VERSION_ERAS,
-  scoreFromAppsResult,
-  scoreFromOAuthResult,
-  scoreFromProtocolResult,
-  scoreFromTasksResult,
   TASKS_CHECK_CATALOG,
   type ConformanceScore,
-  type McpProtocolVersion,
   type OAuthConformanceCheckId,
 } from "@mcpjam/sdk/browser";
-import type { OAuthConformanceStartResult } from "@/lib/apis/mcp-conformance-api";
 import {
-  runProtocolConformance,
-  runAppsConformance,
-  runTasksConformance,
-  startOAuthConformance,
-  submitOAuthConformanceCode,
-  completeOAuthConformance,
-} from "@/lib/apis/mcp-conformance-api";
-import { deriveOAuthProfileFromServer } from "@/components/oauth/utils";
-
-type SuiteStatus = "idle" | "running" | "done" | "error" | "unavailable";
-
-/** "auto" ⇒ no pin: the run adopts whatever version the server negotiates. */
-type ProtocolVersionPin = "auto" | McpProtocolVersion;
+  useConformanceRun,
+  type ProtocolVersionPin,
+  type SuiteState,
+} from "@/hooks/use-conformance-run";
 
 /** One row in a suite's "what this will run" preview, before any run. */
 interface CatalogEntry {
@@ -123,105 +103,6 @@ const OAUTH_CATALOG: CatalogEntry[] = (
   title: CONFORMANCE_CHECK_METADATA[id].title,
   description: CONFORMANCE_CHECK_METADATA[id].summary,
 }));
-
-interface SuiteState {
-  status: SuiteStatus;
-  error?: string;
-  unavailableReason?: string;
-  /**
-   * The verdict of a finished run. "Done" only says the run ended — a suite
-   * that finished with failures must not read as green.
-   */
-  verdict?: "passed" | "failed" | "incomplete";
-}
-
-interface ProtocolSuiteState extends SuiteState {
-  result?: MCPConformanceResult;
-}
-
-interface AppsSuiteState extends SuiteState {
-  result?: MCPAppsConformanceResult;
-}
-
-interface TasksSuiteState extends SuiteState {
-  result?: MCPTasksConformanceResult;
-}
-
-interface OAuthSuiteState extends SuiteState {
-  result?: OAuthConformanceStartResult["result"];
-  sessionId?: string;
-  waitingForAuth?: boolean;
-}
-
-function isHttpServer(server: ServerWithName): boolean {
-  // `server.config` is typed required but can be undefined during hosted
-  // hydration — guard before using the `in` operator to avoid a TypeError.
-  return !!server.config && "url" in server.config;
-}
-
-function suiteState(
-  suite: "protocol" | "oauth" | "apps" | "tasks",
-  server: ServerWithName,
-): SuiteState {
-  // The SDK's `canRunConformance` is the source of truth for which suites
-  // support which transports. It handles null/undefined configs gracefully.
-  const support = canRunConformance(
-    suite,
-    server.config as Parameters<typeof canRunConformance>[1],
-  );
-  return support.supported
-    ? { status: "idle" }
-    : { status: "unavailable", unavailableReason: support.reason };
-}
-
-function createProtocolState(server: ServerWithName): ProtocolSuiteState {
-  return suiteState("protocol", server);
-}
-
-function createAppsState(server: ServerWithName): AppsSuiteState {
-  return suiteState("apps", server);
-}
-
-function createTasksState(server: ServerWithName): TasksSuiteState {
-  return suiteState("tasks", server);
-}
-
-function createOAuthState(server: ServerWithName): OAuthSuiteState {
-  return suiteState("oauth", server);
-}
-
-/**
- * A server that requires no authorization has no OAuth obligations to test —
- * authorization is OPTIONAL in every MCP revision — so the suite reports
- * `not-applicable`. That is not a result to render as red: it reuses the same
- * "unavailable" treatment as a transport that cannot run the suite.
- */
-function oauthStateFromResult(
-  result: NonNullable<OAuthConformanceStartResult["result"]>,
-): OAuthSuiteState {
-  if (result.outcome === "not-applicable") {
-    return {
-      status: "unavailable",
-      unavailableReason: result.summary,
-      // Kept so the score pool can see the run: a not-applicable OAuth run
-      // contributes its "nothing to score here" tally instead of vanishing.
-      result,
-    };
-  }
-  return {
-    status: "done",
-    result,
-    // OAuth now reports the same three-value verdict as the other suites
-    // (plus its own not-applicable, handled above), so an incomplete flow is
-    // badged amber rather than flattened to failed.
-    verdict:
-      result.outcome === "incomplete"
-        ? "incomplete"
-        : result.passed
-          ? "passed"
-          : "failed",
-  };
-}
 
 function StatusIcon({ status }: { status: string }) {
   if (status === "passed") {
@@ -532,9 +413,7 @@ function SuiteSection({
         return <span className="text-[10px] text-red-400">Failed</span>;
       }
       if (state.verdict === "incomplete") {
-        return (
-          <span className="text-[10px] text-amber-500">Incomplete</span>
-        );
+        return <span className="text-[10px] text-amber-500">Incomplete</span>;
       }
       return <span className="text-[10px] text-green-500">Passed</span>;
     }
@@ -605,406 +484,31 @@ function SuiteSection({
 }
 
 function ConformanceContent({ server }: { server: ServerWithName }) {
-  const httpServer = isHttpServer(server);
-  const [protocol, setProtocol] = useState<ProtocolSuiteState>(() =>
-    createProtocolState(server),
-  );
-  const [apps, setApps] = useState<AppsSuiteState>(() =>
-    createAppsState(server),
-  );
-  const [tasks, setTasks] = useState<TasksSuiteState>(() =>
-    createTasksState(server),
-  );
-  const [oauth, setOAuth] = useState<OAuthSuiteState>(() =>
-    createOAuthState(server),
-  );
-  const [versionPin, setVersionPin] = useState<ProtocolVersionPin>("auto");
-  const [runVersion, setRunVersion] = useState(0);
-
-  const activeServerNameRef = useRef(server.name);
-  const latestRunTokenRef = useRef(0);
-  const oauthListenerCleanupRef = useRef<(() => void) | null>(null);
-
-  const clearOAuthListeners = useCallback(() => {
-    oauthListenerCleanupRef.current?.();
-    oauthListenerCleanupRef.current = null;
-  }, []);
-
-  const beginRun = useCallback(() => {
-    latestRunTokenRef.current += 1;
-    clearOAuthListeners();
-    return latestRunTokenRef.current;
-  }, [clearOAuthListeners]);
-
-  const isRunActive = useCallback(
-    (runToken: number, serverName: string) =>
-      latestRunTokenRef.current === runToken &&
-      activeServerNameRef.current === serverName,
-    [],
-  );
-
-  const resetStates = useCallback(
-    (serverName?: string) => {
-      const effectiveServerName = serverName ?? activeServerNameRef.current;
-      latestRunTokenRef.current += 1;
-      clearOAuthListeners();
-      setRunVersion((value) => value + 1);
-      setProtocol(createProtocolState(server));
-      setApps(createAppsState(server));
-      setTasks(createTasksState(server));
-      setOAuth(createOAuthState(server));
-      activeServerNameRef.current = effectiveServerName;
-    },
-    [clearOAuthListeners, server],
-  );
-
-  useEffect(() => {
-    resetStates(server.name);
-  }, [server.name, resetStates]);
-
-  useEffect(
-    () => () => {
-      latestRunTokenRef.current += 1;
-      clearOAuthListeners();
-    },
-    [clearOAuthListeners],
-  );
-
-  const runProtocol = useCallback(
-    async (runToken: number, serverName: string) => {
-      if (!httpServer) return;
-      setProtocol({ status: "running" });
-      try {
-        const { result } = await runProtocolConformance(serverName, {
-          protocolVersion: versionPin === "auto" ? undefined : versionPin,
-        });
-        if (!isRunActive(runToken, serverName)) return;
-        setProtocol({
-          status: "done",
-          result,
-          // All four suites now report a three-value outcome, so an
-          // incomplete run is badged as such rather than flattened to failed.
-          verdict:
-            result.outcome === "incomplete"
-              ? "incomplete"
-              : result.passed
-                ? "passed"
-                : "failed",
-        });
-      } catch (err) {
-        if (!isRunActive(runToken, serverName)) return;
-        setProtocol({
-          status: "error",
-          error: err instanceof Error ? err.message : String(err),
-        });
-      }
-    },
-    [httpServer, isRunActive, versionPin],
-  );
-
-  const runApps = useCallback(
-    async (runToken: number, serverName: string) => {
-      setApps({ status: "running" });
-      try {
-        const { result } = await runAppsConformance(serverName);
-        if (!isRunActive(runToken, serverName)) return;
-        setApps({
-          status: "done",
-          result,
-          // All four suites now report a three-value outcome, so an
-          // incomplete run is badged as such rather than flattened to failed.
-          verdict:
-            result.outcome === "incomplete"
-              ? "incomplete"
-              : result.passed
-                ? "passed"
-                : "failed",
-        });
-      } catch (err) {
-        if (!isRunActive(runToken, serverName)) return;
-        setApps({
-          status: "error",
-          error: err instanceof Error ? err.message : String(err),
-        });
-      }
-    },
-    [isRunActive],
-  );
-
-  const runTasks = useCallback(
-    async (runToken: number, serverName: string) => {
-      setTasks({ status: "running" });
-      try {
-        const { result } = await runTasksConformance(serverName);
-        if (!isRunActive(runToken, serverName)) return;
-        setTasks({
-          status: "done",
-          result,
-          // The tasks suite reports a real third outcome: checks that apply but
-          // could not be exercised leave the run incomplete, never passing.
-          verdict:
-            result.outcome === "incomplete"
-              ? "incomplete"
-              : result.passed
-                ? "passed"
-                : "failed",
-        });
-      } catch (err) {
-        if (!isRunActive(runToken, serverName)) return;
-        setTasks({
-          status: "error",
-          error: err instanceof Error ? err.message : String(err),
-        });
-      }
-    },
-    [isRunActive],
-  );
-
-  const pollOAuthComplete = useCallback(
-    async (sessionId: string, runToken: number, serverName: string) => {
-      const MAX_POLLS = 10;
-      for (let i = 0; i < MAX_POLLS; i++) {
-        try {
-          const poll = await completeOAuthConformance(sessionId);
-          if (!isRunActive(runToken, serverName)) return;
-          if (poll.phase === "complete" && poll.result) {
-            setOAuth(oauthStateFromResult(poll.result));
-            return;
-          }
-        } catch (err) {
-          if (!isRunActive(runToken, serverName)) return;
-          setOAuth({
-            status: "error",
-            error: err instanceof Error ? err.message : String(err),
-          });
-          return;
-        }
-      }
-
-      if (!isRunActive(runToken, serverName)) return;
-      setOAuth({ status: "error", error: "OAuth conformance timed out" });
-    },
-    [isRunActive],
-  );
-
-  const handleOAuthCallback = useCallback(
-    async (
-      sessionId: string,
-      code: string,
-      runToken: number,
-      serverName: string,
-      state?: string,
-    ) => {
-      try {
-        await submitOAuthConformanceCode({ sessionId, code, state });
-        if (!isRunActive(runToken, serverName)) return;
-        setOAuth((prev) => ({
-          ...prev,
-          waitingForAuth: false,
-          status: "running",
-        }));
-        await pollOAuthComplete(sessionId, runToken, serverName);
-      } catch (err) {
-        if (!isRunActive(runToken, serverName)) return;
-        setOAuth({
-          status: "error",
-          error: err instanceof Error ? err.message : String(err),
-        });
-      }
-    },
-    [isRunActive, pollOAuthComplete],
-  );
-
-  const runOAuth = useCallback(
-    async (runToken: number, currentServer: ServerWithName) => {
-      if (!isHttpServer(currentServer)) return;
-
-      const serverName = currentServer.name;
-      setOAuth({ status: "running" });
-
-      try {
-        const profile = deriveOAuthProfileFromServer(currentServer);
-        // Always send callbackOrigin — both local and hosted modes redirect
-        // back to the inspector's own `/oauth/callback/debug` page so the
-        // server can surface the code back to the SDK runner.
-        const callbackOrigin = window.location.origin;
-
-        // A run-scoped version pin beats the stored OAuth-tab profile, and
-        // must be sent even when the profile is otherwise empty (the route
-        // falls back to its own default version when no profile arrives).
-        const pinnedVersion = versionPin === "auto" ? undefined : versionPin;
-        const baseProfile = profile.serverUrl
-          ? {
-              serverUrl: profile.serverUrl,
-              protocolVersion: profile.protocolVersion,
-              registrationStrategy: profile.registrationStrategy,
-              clientId: profile.clientId || undefined,
-              clientSecret: profile.clientSecret || undefined,
-              scopes: profile.scopes || undefined,
-              customHeaders: profile.customHeaders.length
-                ? profile.customHeaders
-                : undefined,
-            }
-          : undefined;
-
-        const startResult = await startOAuthConformance({
-          serverNameOrId: serverName,
-          oauthProfile: pinnedVersion
-            ? { ...baseProfile, protocolVersion: pinnedVersion }
-            : baseProfile,
-          callbackOrigin,
-        });
-
-        if (!isRunActive(runToken, serverName)) return;
-
-        if (startResult.phase === "complete" && startResult.result) {
-          setOAuth(oauthStateFromResult(startResult.result));
-          return;
-        }
-
-        if (
-          startResult.phase === "authorization_needed" &&
-          startResult.sessionId &&
-          startResult.authorizationUrl
-        ) {
-          const sessionId = startResult.sessionId;
-          setOAuth({
-            status: "running",
-            sessionId,
-            waitingForAuth: true,
-          });
-
-          window.open(
-            startResult.authorizationUrl,
-            "oauth_conformance_auth",
-            "width=600,height=700,scrollbars=yes",
-          );
-
-          const cleanupFns: Array<() => void> = [];
-          const cleanup = () => {
-            for (const fn of cleanupFns) fn();
-            if (oauthListenerCleanupRef.current === cleanup) {
-              oauthListenerCleanupRef.current = null;
-            }
-          };
-          oauthListenerCleanupRef.current = cleanup;
-
-          const handleMessage = (event: MessageEvent) => {
-            if (event.data?.type !== "OAUTH_CALLBACK" || !event.data?.code) {
-              return;
-            }
-            cleanup();
-            void handleOAuthCallback(
-              sessionId,
-              event.data.code,
-              runToken,
-              serverName,
-              event.data.state,
-            );
-          };
-
-          window.addEventListener("message", handleMessage);
-          cleanupFns.push(() =>
-            window.removeEventListener("message", handleMessage),
-          );
-
-          try {
-            const channel = new BroadcastChannel("oauth_callback_channel");
-            channel.onmessage = (event) => {
-              if (event.data?.type !== "OAUTH_CALLBACK" || !event.data?.code) {
-                return;
-              }
-              cleanup();
-              void handleOAuthCallback(
-                sessionId,
-                event.data.code,
-                runToken,
-                serverName,
-                event.data.state,
-              );
-            };
-            cleanupFns.push(() => channel.close());
-          } catch {
-            // BroadcastChannel not available
-          }
-
-          // Both local and hosted modes now rely on the `/oauth/authorize`
-          // endpoint to deliver the code; polling without a code submission
-          // would time out, so we defer all polling to `handleOAuthCallback`.
-        }
-      } catch (err) {
-        if (!isRunActive(runToken, currentServer.name)) return;
-        setOAuth({
-          status: "error",
-          error: err instanceof Error ? err.message : String(err),
-        });
-      }
-    },
-    [handleOAuthCallback, isRunActive, pollOAuthComplete, versionPin],
-  );
-
-  const runAll = useCallback(async () => {
-    const runToken = beginRun();
-    const currentServer = server;
-    const httpServerNow = isHttpServer(currentServer);
-
-    setRunVersion((value) => value + 1);
-    setProtocol(createProtocolState(currentServer));
-    setApps(createAppsState(currentServer));
-    setTasks(createTasksState(currentServer));
-    setOAuth(createOAuthState(currentServer));
-
-    const promises: Promise<void>[] = [];
-    if (httpServerNow) {
-      promises.push(runProtocol(runToken, currentServer.name));
-    }
-    promises.push(runApps(runToken, currentServer.name));
-    promises.push(runTasks(runToken, currentServer.name));
-    if (httpServerNow) {
-      promises.push(runOAuth(runToken, currentServer));
-    }
-
-    await Promise.allSettled(promises);
-  }, [beginRun, runApps, runOAuth, runProtocol, runTasks, server]);
+  // Run state, per-suite scores and the pooled headline all live in the shared
+  // hook — score.mcpjam.com runs the same four suites and must pool them the
+  // same way. Rendering stays here.
+  const {
+    protocol,
+    apps,
+    tasks,
+    oauth,
+    versionPin,
+    setVersionPin,
+    runVersion,
+    runAll,
+    protocolScore,
+    appsScore,
+    tasksScore,
+    oauthScore,
+    pooledScore,
+    oauthNotScored,
+    isRunning,
+  } = useConformanceRun({ server });
 
   const protocolChecks = useMemo(
     () => protocolCatalog(versionPin),
-    [versionPin],
+    [versionPin]
   );
-
-  // Scores are derived, never stored: each finished suite contributes, and
-  // the headline pools their COUNTS (not an average of suite scores), so a
-  // 0-for-9 OAuth run stays visible inside 30 protocol passes.
-  const protocolScore = useMemo(
-    () => (protocol.result ? scoreFromProtocolResult(protocol.result) : undefined),
-    [protocol.result],
-  );
-  const appsScore = useMemo(
-    () => (apps.result ? scoreFromAppsResult(apps.result) : undefined),
-    [apps.result],
-  );
-  const tasksScore = useMemo(
-    () => (tasks.result ? scoreFromTasksResult(tasks.result) : undefined),
-    [tasks.result],
-  );
-  const oauthScore = useMemo(
-    () => (oauth.result ? scoreFromOAuthResult(oauth.result) : undefined),
-    [oauth.result],
-  );
-  const pooledScore = useMemo(() => {
-    const parts = [protocolScore, appsScore, tasksScore, oauthScore].filter(
-      (part): part is ConformanceScore => part !== undefined,
-    );
-    return parts.length > 0 ? pooledConformanceScore(parts) : undefined;
-  }, [protocolScore, appsScore, tasksScore, oauthScore]);
-  const oauthNotScored =
-    oauthScore !== undefined && oauthScore.score === null;
-
-  const isRunning =
-    protocol.status === "running" ||
-    apps.status === "running" ||
-    tasks.status === "running" ||
-    oauth.status === "running";
 
   return (
     <div className="flex h-full flex-col overflow-hidden">
@@ -1031,17 +535,17 @@ function ConformanceContent({ server }: { server: ServerWithName }) {
                 pooledScore.outcome === "failed"
                   ? "text-red-400"
                   : pooledScore.outcome === "incomplete"
-                    ? "text-amber-500"
-                    : "text-green-500"
+                  ? "text-amber-500"
+                  : "text-green-500"
               }`}
             >
               {pooledScore.outcome === "failed"
                 ? "Not conformant"
                 : pooledScore.outcome === "incomplete"
-                  ? "Incomplete run"
-                  : pooledScore.advisories.length > 0
-                    ? "Conformant, with advice"
-                    : "Fully conformant"}
+                ? "Incomplete run"
+                : pooledScore.advisories.length > 0
+                ? "Conformant, with advice"
+                : "Fully conformant"}
             </div>
             {/* The denominator and version travel with the number, always —
                 "100 of 11 applicable" and "100 of 38" are different servers. */}
@@ -1050,7 +554,9 @@ function ConformanceContent({ server }: { server: ServerWithName }) {
               {pooledScore.notApplicable > 0
                 ? ` · ${pooledScore.notApplicable} not applicable`
                 : ""}
-              {oauthNotScored ? " · OAuth not applicable (no auth) — not scored" : ""}
+              {oauthNotScored
+                ? " · OAuth not applicable (no auth) — not scored"
+                : ""}
             </div>
           </div>
         </div>
@@ -1076,7 +582,9 @@ function ConformanceContent({ server }: { server: ServerWithName }) {
           </label>
           <Select
             value={versionPin}
-            onValueChange={(value) => setVersionPin(value as ProtocolVersionPin)}
+            onValueChange={(value) =>
+              setVersionPin(value as ProtocolVersionPin)
+            }
             disabled={isRunning}
           >
             <SelectTrigger
@@ -1212,11 +720,7 @@ function ConformanceContent({ server }: { server: ServerWithName }) {
   );
 }
 
-export function ConformanceTab({
-  server,
-}: {
-  server?: ServerWithName | null;
-}) {
+export function ConformanceTab({ server }: { server?: ServerWithName | null }) {
   // In hosted mode `selectedMCPConfig` can arrive as a stub with falsy name
   // and/or missing config while the project is still hydrating — treat any
   // non-connected shape as "no server selected" so the panel never runs
